# backend/customer-loader-backend/src/main/resources/META-INF/openapi.yaml
openapi: 3.0.3
info:
  title: Customer Loader API
  version: 1.0.0
  description: API para carga masiva de clientes y gestión de nómina
servers:
  - url: http://localhost:8081/customer-loader-backend/api
    description: Development

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /clients:
    get:
      tags:
        - Clients
      summary: Get all clients
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of clients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse'

  /clients/{id}:
    get:
      tags:
        - Clients
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Client details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          description: Client not found

  /accounts/client/{clientId}:
    get:
      tags:
        - Accounts
      summary: Get account by client ID
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Account not found

  /accounts/client/{accountId}/first-payment:
    get:
      tags:
        - Accounts
      summary: Get first payment by account ID
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayrollPayment'
        '404':
          description: Payment not found

  /bulk-load:
    post:
      tags:
        - Bulk Load
      summary: Upload clients file
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkLoadResponse'
        '400':
          description: Invalid file format

  /processes:
    get:
      tags:
        - Processes
      summary: Get all processes
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of processes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse'

  /processes/{id}:
    get:
      tags:
        - Processes
      summary: Get process by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Process details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkLoadProcess'

  /errors:
    get:
      tags:
        - Errors
      summary: Get all errors
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse'

components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: "UP"
        timestamp:
          type: string
          format: date-time

    Client:
      type: object
      properties:
        id:
          type: integer
          format: int64
        idType:
          type: string
          enum: [C, P]
        idNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        birthDate:
          type: string
          format: date
        payrollValue:
          type: number
          format: double

    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
        accountNumber:
          type: string
        balance:
          type: number
          format: double
        status:
          type: string
        clientId:
          type: integer
          format: int64

    PayrollPayment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        accountId:
          type: integer
          format: int64
        paymentDate:
          type: string
          format: date
        amount:
          type: number
          format: double
        status:
          type: string

    BulkLoadProcess:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fileName:
          type: string
        status:
          type: string
        totalRecords:
          type: integer
        processedRecords:
          type: integer
        errorRecords:
          type: integer
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time

    BulkLoadResponse:
      type: object
      properties:
        processId:
          type: integer
          format: int64
        message:
          type: string
        status:
          type: string

    PagedResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string